requirements.txt
streamlit>=1.22
pandas>=2.0
requests>=2.28
plotly>=5.15
matplotlib>=3.7
python-dotenv>=1.0
sqlalchemy>=2.0
pydantic>=1.10
rapidfuzz>=2.15